requirements/base.txt
Django==1.9.4
pytz==2015.7
django-widget-tweaks==1.4.1
psycopg2==2.6.1
django-libsass==0.6
django_compressor==2.0
python-dateutil==2.5.3
whoosh==2.7.4
git+git://github.com/django-haystack/django-haystack.git
djangorestframework
django-webpack-loader
requests